// services/api/src/modules/school/data/models/pastQuestions.ts
import { Media, PastQuestionData } from '../../domain/types'

export interface PastQuestionFromModel extends PastQuestionToModel {
	_id: string
	createdAt: number
	updatedAt: number
}

export interface PastQuestionToModel {
	institutionId: string
	courseId: string
	year: number
	question: string
	questionMedia: Media[]
	data: PastQuestionData
}